Organization & make sure we don't add already queued repos to queue
